refactor(slides): collapse the pipeline into one module
Twenty .mjs files for 1160 lines meant most of them existed to re-import each
other, and the split had started to lie: the shared acceptance rules lived in
validate-slides.mjs because that is who enforces them, while collect-candidates
and caption-agent imported them to screen their own output.

One file, sectioned in dependency order, drops that indirection and the three
separate CLI entrypoints with it. slides.js collect|refresh|validate is now the
only way in, so package.json, pr-test.yml and the refresh workflow all name the
same command.

No behaviour change: the same 52 tests pass and slides.json validates unchanged.
