+++ package.json
@@ -16,9 +16,10 @@
     "postbuild": "pagefind --site dist",
     "test:slugs": "node scripts/check-slugs.mjs",
     "test:pages": "node scripts/test-pages.mjs",
-    "slides:collect": "node scripts/slides/collect-candidates.mjs",
-    "slides:refresh": "node scripts/slides/refresh.mjs",
-    "slides:validate": "node scripts/slides/validate-slides.mjs"
+    "slides:collect": "node scripts/slides/slides.js collect",
+    "slides:refresh": "node scripts/slides/slides.js refresh",
+    "slides:validate": "node scripts/slides/slides.js validate",
+    "slides:test": "node --test scripts/slides/slides.test.js"
   },
   "dependencies": {
     "@astrojs/check": "^0.9.0",
